RR-48 reorganize code in preparation for vision test app
