+++ RobotLib/build.gradle
@@ -28,6 +28,8 @@ dependencies {
     implementation(name: 'Analytics-release', ext: 'aar')
     implementation(name: 'WirelessP2p-release', ext: 'aar')
 
+    implementation project(':openCVLibrary331')
+
     // already in a jar
 //    implementation 'org.nanohttpd:nanohttpd:2.3.1'
 //    implementation('org.nanohttpd:nanohttpd-webserver:2.3.1') {
