+++ RobotLib/src/main/java/com/acmerobotics/relicrecovery/configuration/Cryptobox.java
@@ -1,4 +1,4 @@
-package com.acmerobotics.library.configuration;
+package com.acmerobotics.relicrecovery.configuration;
 
 import com.acmerobotics.library.localization.Pose2d;
 
@@ -6,17 +6,17 @@ import com.acmerobotics.library.localization.Pose2d;
  * Created by ryanbrott on 11/12/17.
  */
 
-public enum BalancingStone {
-    NEAR_BLUE(0, AllianceColor.BLUE, new Pose2d(48, -48, Math.PI)),
-    FAR_BLUE(1, AllianceColor.BLUE, new Pose2d(-24, -48, Math.PI)),
-    NEAR_RED(2, AllianceColor.RED, new Pose2d(48, 48, 0)),
-    FAR_RED(3, AllianceColor.RED, new Pose2d(-24, 48, 0));
+public enum Cryptobox {
+    NEAR_BLUE(0, AllianceColor.BLUE, new Pose2d(12, -72, Math.PI / 2)),
+    FAR_BLUE(1, AllianceColor.BLUE, new Pose2d(-72, -36, 0)),
+    NEAR_RED(2, AllianceColor.RED, new Pose2d(12, 72, - Math.PI / 2)),
+    FAR_RED(3, AllianceColor.RED, new Pose2d(-72, 36, 0));
 
     private int index;
     private AllianceColor allianceColor;
     private Pose2d pose;
 
-    BalancingStone(int i, AllianceColor color, Pose2d pose) {
+    Cryptobox(int i, AllianceColor color, Pose2d pose) {
         this.index = i;
         this.allianceColor = color;
         this.pose = pose;
@@ -26,8 +26,8 @@ public enum BalancingStone {
         return index;
     }
 
-    public static BalancingStone fromIndex(int i) {
-        for (BalancingStone color : BalancingStone.values()) {
+    public static Cryptobox fromIndex(int i) {
+        for (Cryptobox color : Cryptobox.values()) {
             if (color.getIndex() == i) {
                 return color;
             }
@@ -39,11 +39,11 @@ public enum BalancingStone {
         return allianceColor;
     }
 
-    public Cryptobox getCryptobox() {
-        return Cryptobox.fromIndex(index);
-    }
-
     public Pose2d getPose() {
         return pose;
+    }
+
+    public BalancingStone getBalancingStone() {
+        return BalancingStone.fromIndex(index);
     }
 }
