+++ RobotLib/src/main/res/drawable/vumark.png
