+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/drive/CryptoboxLocalizer.java
@@ -1,13 +1,14 @@
-package com.acmerobotics.relicrecovery.vision;
+package com.acmerobotics.relicrecovery.drive;
 
 import android.util.Log;
 
-import com.acmerobotics.library.configuration.AllianceColor;
-import com.acmerobotics.library.configuration.Cryptobox;
 import com.acmerobotics.library.dashboard.config.Config;
 import com.acmerobotics.library.localization.Pose2d;
 import com.acmerobotics.library.localization.Vector2d;
-import com.acmerobotics.relicrecovery.drive.MecanumDrive;
+import com.acmerobotics.relicrecovery.configuration.AllianceColor;
+import com.acmerobotics.relicrecovery.configuration.Cryptobox;
+import com.acmerobotics.relicrecovery.vision.CryptoboxTracker;
+import com.acmerobotics.relicrecovery.vision.VisionCamera;
 
 import java.util.ArrayList;
 import java.util.Collections;
