+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/drive/PathFollower.java
@@ -3,6 +3,7 @@ package com.acmerobotics.relicrecovery.drive;
 import com.acmerobotics.library.dashboard.config.Config;
 import com.acmerobotics.library.localization.Pose2d;
 import com.acmerobotics.library.localization.Vector2d;
+import com.acmerobotics.library.util.TimestampedData;
 import com.acmerobotics.relicrecovery.motion.MotionState;
 import com.acmerobotics.relicrecovery.motion.PIDFCoefficients;
 import com.acmerobotics.relicrecovery.motion.PIDFController;
