+++ TeamCode/src/test/java/com/acmerobotics/relicrecovery/path/PathTest.java
@@ -3,7 +3,7 @@ package com.acmerobotics.relicrecovery.path;
 import com.acmerobotics.library.localization.Pose2d;
 import com.acmerobotics.library.localization.Vector2d;
 import com.acmerobotics.relicrecovery.drive.PathFollower;
-import com.acmerobotics.relicrecovery.drive.TimestampedData;
+import com.acmerobotics.library.util.TimestampedData;
 import com.acmerobotics.relicrecovery.motion.PIDFCoefficients;
 
 import org.junit.Ignore;
