+++ TeamCode/src/test/java/com/acmerobotics/relicrecovery/vision/CryptoboxTrackerTest.java
@@ -1,7 +1,6 @@
 package com.acmerobotics.relicrecovery.vision;
 
-import com.acmerobotics.library.configuration.AllianceColor;
-import com.acmerobotics.relicrecovery.util.VisionUtil;
+import com.acmerobotics.relicrecovery.configuration.AllianceColor;
 
 import org.junit.Ignore;
 import org.junit.Test;
